# perf(rome_cli): replace global allocator with jemalloc