--- a/crates/rome_cli/Cargo.toml
+++ b/crates/rome_cli/Cargo.toml
@@ -42,6 +42,9 @@ tokio = { version = "1.15.0", features = ["process"] }
 [target.'cfg(windows)'.dependencies]
 mimalloc = "0.1.29"
 
+[target.'cfg(target_os = "macos")'.dependencies]
+tikv-jemallocator = "0.5.0"
+
 [dev-dependencies]
 insta = "1.18.2"
 tokio = { version = "1.15.0", features = ["io-util"] }