--- a/crates/rome_cli/Cargo.toml
+++ b/crates/rome_cli/Cargo.toml
@@ -42,7 +42,7 @@ tokio = { version = "1.15.0", features = ["process"] }
 [target.'cfg(windows)'.dependencies]
 mimalloc = "0.1.29"
 
-[target.'cfg(target_os = "macos")'.dependencies]
+[target.'cfg(not(windows))'.dependencies]
 tikv-jemallocator = "0.5.0"
 
 [dev-dependencies]