--- a/xtask/bench/Cargo.toml
+++ b/xtask/bench/Cargo.toml
@@ -30,5 +30,8 @@ humansize = {version = "1.1.1", optional = true }
 [target.'cfg(target_os = "windows")'.dependencies]
 mimalloc = "0.1.29"
 
+[target.'cfg(target_os = "macos")'.dependencies]
+tikv-jemallocator = "0.5.0"
+
 [features]
 dhat-heap = ["dhat", "humansize"]