@MichaReiser 2022-09-16
Have you compared performance with mimalloc? Depending on different allocators for every platform increases the risk of platforms specific issues and requires us to test every change on every platform. 

That's why I would prefer to use the default system allocator for all (maybe except windows) platforms or replace the default for all platforms.