@MichaReiser 2022-09-16
You know... this will make it more difficult for me to prove that any of my work reducing allocation is any good :P