@IWANABETHATGUY 2022-09-16
> 

Would you mind explaining why replacing an allocator will make it harder to prove reducing allocation ?