@IWANABETHATGUY 2022-09-16
> > > 
> > 
> > 
> > Would you mind explaining why replacing an allocator will make it harder to prove reducing allocation ?
> 
> If every single allocations becomes cheaper than reducing the number of allocations has less effect on the overall performance.
> 
> @ematipico Would you mind also comparing the memory consumption? Do you have links that compare the different allocators?

Thanks