@MichaReiser 2022-09-16
> > Depending on different allocators for every platform increases the risk of platforms specific issues and requires us to test every change on every platform.
> 
> Technically we are already depending on different allocators for different platforms, since we use whatever allocator is provided by the OS. I don't know if `mimalloc` compares favorably to `jemalloc` in terms of performance, as always with perfs this is something we need to benchmark with our workloads to be sure. In any case I think it would be interesting to use `jemalloc` on all Unix systems instead of just macOS and `mimalloc` on Windows, this way we would still smooth out most platform discrepancies by using the same allocator (almost) everywhere
> 
> Also, what's the impact of this change on the size of the `rome` binary ?

What was your reasoning why you preferred `minmalloc` over `jemalloc`?