@ematipico 2022-09-16
The reason why I tried `jemalloctor` for macOS first was mainly because [`lightingcss` uses it](https://github.com/parcel-bundler/lightningcss/blob/4e36e076766dd21eaca2cc97fd3b1ed97dd60fb6/Cargo.toml#L48) (former ParcelCSS). And I gave it a try and see what's the result on our benches.

I will give it a whirl in terms of memory usages once I manage to download again all the necessary software. We can close it if you think it's not worth it.