@leops 2022-09-16
> What was your reasoning why you preferred `minmalloc` over `jemalloc`?

For Windows the reason was simply that `tikv-jemallocator` doesn't support it