@ematipico 2022-09-20
So, should we just switch to use this allocator for all the OS expect windows?