@MichaReiser 2022-09-20
> So, should we just switch to use this allocator for all the OS expect windows?

Probably. But it would still be interesting to get some more numbers:

* How does it change performance results on linux
* Is there a way for us to compare memory consumption before/after? If so, what are the numbers