@MichaReiser 2022-09-21
> formatter/typescript.js                  1.00       2.1±0.03s     4.5 MB/sec

Uhm, what happened with typescript?