@ematipico 2022-09-21
> > formatter/typescript.js                  1.00       2.1±0.03s     4.5 MB/sec
> 
> Uhm, what happened with typescript?

Something went wrong during the download of the file:
> "Status(520, Response[status: 520, status_text: , url: https://www.unpkg.com/typescript@4.5.5/lib/typescript.js])"