@ematipico 2022-09-21
@leops and @xunilrj what do you think? The numbers are promises and so far I could not see any issue locally. Should we make the switch to `jemalloc` for all operative systems? (exception for windows)